requirements.txt
beautifulsoup4==4.14.3
soupsieve==2.8
typing_extensions==4.15.0

python-dotenv~=1.2.1